format and move to helper namespace
